[GAFF] Added clearNoFree() to Map.
[GLEAM]
Fixed compilation warnings in Windows' Window.

[SHIBBOLETH]
Added CameraComponent to ComponentsModule.
Added CameraComponent.schema.
Fixed crash bug in RenderTargetLoader when creating DepthStencil buffer.
Fixed crash bugs in OcclusionManager when querying for objects in frustum when the tree was empty or only had one node.
FrameManager now binds camera render target and executes the stored command lists.
InGameRenderPipeline should be generating command lists now.

Test scene now consists of a mesh object and a camera object.

+++ dependencies/Gleam/Gleam_Window_Windows.cpp
@@ -334,7 +334,8 @@ void Window::containCursor(bool contain)
 	if (contain) {
 		RECT rect = {
 			_pos_x, _pos_y,
-			_pos_x + _width, _pos_y + _height
+			_pos_x + static_cast<int>(_width),
+			_pos_y + static_cast<int>(_height)
 		};
 
 		ClipCursor(&rect);
